backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic[email]==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.5
pytest==7.3.1
httpx==0.24.1
aiohttp==3.9.1
websockets==12.0
pandas==2.1.3
numpy==1.26.2
ta==0.10.2
python-binance==1.0.17
psycopg2-binary==2.9.9
alembic==1.12.1
tenacity==8.2.3
python-dateutil==2.8.2
pytz==2023.3.post1
python-socketio==5.10.0 